requirements.txt
# SteamLink Store
PyYAML>=3.11,<4
aiohttp>=3.0.0
aiomqtt>=0.1.0
python-socketio>=1.8.4
paho-mqtt>=1.3.1,<2
hbmqtt>=0.9.1
jinja2==2.10
aiohttp-jinja2>=1.0.0
tinydb>=3.9.0
tinydb-smartcache>=1.0.2
